requirements.txt
# FastAPI and related
fastapi==0.115.9
uvicorn[standard]==0.30.6
pydantic==2.9.2
pydantic[email]==2.9.2
email-validator==2.0.0
python-dotenv==1.0.1

# Firebase
firebase-admin==6.9.0

# AI/ML libraries
crewai==0.130.0
langchain==0.2.14
langchain-openai==0.1.25
langchain-core==0.2.43
langgraph==0.4.5
openai==1.88.0

# Google APIs
google-api-python-client==2.172.0
google-auth==2.34.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
google-cloud-storage==3.1.0

# PDF processing
pdfplumber==0.10.4
pypdf==4.1.0
pdf2image==1.17.0
reportlab==4.0.8  # For generating PDF documents

# Utilities
requests==2.32.3
pytest==8.2.6
httpx==0.28.1
uuid==1.30
